+++ packages/components/src/components/CartesianChart/stories/Default.stories.tsx
@@ -9,6 +9,7 @@ import CartesianGrid from "@/components/CartesianChart/components/CartesianGrid/
 import YAxis from "@/components/CartesianChart/components/YAxis/YAxis";
 import XAxis from "@/components/CartesianChart/components/XAxis/XAxis";
 import ChartLegend from "@/components/CartesianChart/components/ChartLegend/ChartLegend";
+import { sleep } from "@/lib/promises/sleep";
 
 const chartData = [
   {
@@ -172,11 +173,19 @@ export const WithLine: Story = {
   render: (props) => (
     <CartesianChart {...props} height="70vh">
       <CartesianGrid />
-      <Area dataKey="mean" />
-      <Line dataKey="max" color="magenta" />
+      <Area dataKey="mean" unit="ms" />
+      <Line dataKey="max" color="magenta" unit="ms" />
       <XAxis dataKey="time" />
       <YAxis interval="equidistantPreserveStart" unit="%" domain={[0, 100]} />
-      <ChartTooltip />
+      <ChartTooltip
+        headingFormatter={(v) => {
+          return `Sync Format: ${v}`;
+        }}
+        formatter={async (value, name, index, unit) => {
+          await sleep(3000);
+          return `Async Format: ${name}: ${value}${unit ? ` ${unit}` : ""}`;
+        }}
+      />
       <ChartLegend />
     </CartesianChart>
   ),